requirements.txt
fastapi
uvicorn
aiohttp
python-dotenv
aiogram==3.*
